🎈 perf: 优化错误处理与graphql type bigint字段类型错误

+++ src/Console/GraphQLTypeMakeCommand.php
@@ -47,7 +47,7 @@ class GraphQLTypeMakeCommand extends GeneratorCommand
     protected $filterMaps = [
         'integer'  => 'int',
         'smallint' => 'int',
-        'bigint'   => 'int',
+        'bigint'   => 'bigint',
         'string'   => 'string',
         'text'     => 'string',
     ];
